# src/api/target/classes/openapi.yaml
openapi: 3.0.3
info:
  title: INQCUST Customer Inquiry API
  version: 1.0.0
  description: Spring Boot API that modernizes the INQCUST customer inquiry capability using mocked data.
paths:
  /api/v1/customers/{sortCode}/{customerNumber}:
    get:
      summary: Retrieve customer inquiry response
      operationId: getCustomerInquiry
      parameters:
        - name: sortCode
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]{6}$'
            example: '123456'
        - name: customerNumber
          in: path
          required: true
          description: Use 0000000000 for random lookup and 9999999999 for latest lookup.
          schema:
            type: string
            pattern: '^[0-9]{10}$'
            example: '0000000001'
      responses:
        '200':
          description: Customer inquiry completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerInquiryResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found or special lookup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerInquiryResponse'
        '500':
          description: System error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    CustomerInquiryResponse:
      type: object
      required: [legacyStatus, lookupMode]
      properties:
        legacyStatus:
          $ref: '#/components/schemas/LegacyInquiryStatus'
        lookupMode:
          type: string
          enum: [SPECIFIC, RANDOM, LATEST]
        customer:
          $ref: '#/components/schemas/CustomerResponse'
          nullable: true
        riskAssessment:
          $ref: '#/components/schemas/RiskAssessmentResponse'
          nullable: true
    LegacyInquiryStatus:
      type: object
      required: [inquirySuccess, inquiryFailCode, message]
      properties:
        inquirySuccess:
          type: string
          enum: [Y, N]
        inquiryFailCode:
          type: string
          enum: ['0', '1', '2', '9']
        message:
          type: string
    CustomerResponse:
      type: object
      properties:
        eyecatcher:
          type: string
          maxLength: 4
        sortCode:
          type: string
          pattern: '^[0-9]{6}$'
        customerNumber:
          type: string
          pattern: '^[0-9]{10}$'
        title:
          type: string
          maxLength: 10
        firstName:
          type: string
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        dateOfBirth:
          type: string
          format: date
        phone:
          type: string
          maxLength: 20
        address:
          $ref: '#/components/schemas/AddressResponse'
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
        createdDate:
          type: string
          format: date
        creditScore:
          type: integer
          minimum: 0
          maximum: 999
        creditScoreReviewDate:
          type: string
          format: date
    AddressResponse:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        postcode:
          type: string
        country:
          type: string
    RiskAssessmentResponse:
      type: object
      properties:
        riskRating:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        reviewRequired:
          type: boolean
        reasons:
          type: array
          items:
            type: string
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorResponse'
    FieldErrorResponse:
      type: object
      properties:
        field:
          type: string
        message:
          type: string